+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.1.0]
+- New command 'Fold All keep Cursor Line'
+- Word highlighting now restored after fold
+- Fixed issues with tabs
+
 ## [1.0.2]
 - Word under cursor matches only exact word now.  Selections match any line containing.
 - Fixed issues with emply lines and spaces